# docs/swagger.yaml
basePath: /api
definitions:
  api.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      code:
        description: Error code for client-side error handling
        type: string
      details:
        description: Additional details for debugging
        type: string
      error:
        description: User-friendly error message
        type: string
      request_id:
        description: Request ID for tracing
        type: string
    type: object
...
  model.EventAddress:
    properties:
      address:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.User'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      label:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.User'
      updated_by_id:
        type: integer
    type: object
  model.MealEvent:
    properties:
      addresses:
        items:
          $ref: '#/definitions/model.MealEventAddress'
        type: array
      confirmed_at:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.User'
      created_by_id:
        type: integer
      cutoff_time:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      event_date:
        type: string
      event_duration:
        description: in minutes
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      meal_requests:
        items:
          $ref: '#/definitions/model.MealRequest'
        type: array
      menu_item_comments:
        items:
          $ref: '#/definitions/model.MenuItemComment'
        type: array
      menu_sets:
        items:
          $ref: '#/definitions/model.MealEventSet'
        type: array
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.User'
      updated_by_id:
        type: integer
    type: object
  model.MealEventAddress:
    properties:
      address:
        $ref: '#/definitions/model.EventAddress'
      address_id:
        type: integer
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.User'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      meal_event:
        $ref: '#/definitions/model.MealEvent'
      meal_event_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.User'
      updated_by_id:
        type: integer
    type: object
  model.MealEventSet:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.User'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      label:
        type: string
      meal_event:
        $ref: '#/definitions/model.MealEvent'
      meal_event_id:
        type: integer
      menu_set:
        $ref: '#/definitions/model.MenuSet'
      menu_set_id:
        type: integer
      note:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.User'
      updated_by_id:
        type: integer
    type: object
  model.MealRequest:
    properties:
      confirmed_at:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.User'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      event_address:
        $ref: '#/definitions/model.EventAddress'
      event_address_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      meal_event:
        $ref: '#/definitions/model.MealEvent'
      meal_event_id:
        type: integer
      menu_set:
        $ref: '#/definitions/model.MenuSet'
      menu_set_id:
        type: integer
      request_items:
        items:
          $ref: '#/definitions/model.MealRequestItem'
        type: array
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.User'
      updated_by_id:
        type: integer
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.MealRequestItem:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.User'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_selected:
        type: boolean
      meal_request:
        $ref: '#/definitions/model.MealRequest'
      meal_request_id:
        type: integer
      menu_item:
        $ref: '#/definitions/model.MenuItem'
      menu_item_id:
        type: integer
      menu_set:
        $ref: '#/definitions/model.MenuSet'
      menu_set_id:
        type: integer
      notes:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.User'
      updated_by_id:
        type: integer
    type: object
  model.MenuItem:
    properties:
      average_rating:
        type: number
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.User'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_active:
        type: boolean
      meal_request_items:
        items:
          $ref: '#/definitions/model.MealRequestItem'
        type: array
      menu_item_comments:
        items:
          $ref: '#/definitions/model.MenuItemComment'
        type: array
      menu_set_items:
        items:
          $ref: '#/definitions/model.MenuSetItem'
        type: array
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.User'
      updated_by_id:
        type: integer
    type: object
  model.MenuItemComment:
    properties:
      comment:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.User'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      meal_event:
        $ref: '#/definitions/model.MealEvent'
      meal_event_id:
        type: integer
      menu_item:
        $ref: '#/definitions/model.MenuItem'
      menu_item_id:
        type: integer
      parent:
        $ref: '#/definitions/model.MenuItemComment'
      parent_id:
        description: Added to support replies
        type: integer
      rating:
        type: integer
      replies:
        items:
          $ref: '#/definitions/model.MenuItemComment'
        type: array
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.User'
      updated_by_id:
        type: integer
      user_id:
        type: integer
    type: object
  model.MenuSet:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.User'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      menu_set_description:
        type: string
      menu_set_items:
        items:
          $ref: '#/definitions/model.MenuSetItem'
        type: array
      menu_set_name:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.User'
      updated_by_id:
        type: integer
    type: object
  model.MenuSetItem:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.User'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      menu_item:
        $ref: '#/definitions/model.MenuItem'
      menu_item_id:
        type: integer
      menu_set_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.User'
      updated_by_id:
        type: integer
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.User'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      department:
        type: string
      email:
        type: string
      employee_id:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_login_at:
        type: string
      name:
        type: string
      notification_enabled:
        type: boolean
      role:
        $ref: '#/definitions/model.UserRole'
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.User'
      updated_by_id:
        type: integer
      username:
        type: string
    type: object
  model.UserRole:
    enum:
    - admin
    - employee
    - manager
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleEmployee
    - UserRoleManager
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A meal management system API for employees
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MealSync API
  version: "1.0"
paths:
...
  /menu-sets:
    get:
      consumes:
      - application/json
      description: Get all menu sets for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MenuSet'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List menu sets
      tags:
      - menu-sets
    post:
      consumes:
      - application/json
      description: Create a new menu set
      parameters:
      - description: Menu Set Data
        in: body
        name: menuSet
        required: true
        schema:
          $ref: '#/definitions/model.MenuSet'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.MenuSet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create menu set
      tags:
      - menu-sets
  /menu-sets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing menu set
      parameters:
      - description: Menu Set ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete menu set
      tags:
      - menu-sets
    get:
      consumes:
      - application/json
      description: Get a specific menu set by its ID
      parameters:
      - description: Menu Set ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MenuSet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get menu set by ID
      tags:
      - menu-sets
    put:
      consumes:
      - application/json
      description: Update an existing menu set
      parameters:
      - description: Menu Set ID
        in: path
        name: id
        required: true
        type: integer
      - description: Menu Set Data
        in: body
        name: menuSet
        required: true
        schema:
          $ref: '#/definitions/model.MenuSet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MenuSet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update menu set
      tags:
      - menu-sets
  /menu-sets/{id}/items:
    get:
      consumes:
      - application/json
      description: Get all items in a specific menu set
      parameters:
      - description: Menu Set ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MenuItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List menu set items
      tags:
      - menu-sets
    post:
      consumes:
      - application/json
      description: Add a menu item to an existing menu set
      parameters:
      - description: Menu Set ID
        in: path
        name: id
        required: true
        type: integer
      - description: Menu Item Data
        in: body
        name: menuItem
        required: true
        schema:
          $ref: '#/definitions/model.MenuSetItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.MenuSetItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add menu item to set
      tags:
      - menu-sets
  /menu-sets/{id}/items/{item_id}:
    delete:
      consumes:
      - application/json
      description: Remove a menu item from an existing menu set
      parameters:
      - description: Menu Set ID
        in: path
        name: id
        required: true
        type: integer
      - description: Menu Item ID
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove menu item from set
      tags:
      - menu-sets
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"